Tests for diminishing step size + some sample step sizes

+++ src/optimization_routines/step_size_util/inverse_k.jl
@@ -4,7 +4,7 @@
 
 """
 """
-function inverse_k(c :: Int64)
+function inverse_k(c :: Float64)
     function step_size(k :: Int64)
         return c / k
     end
